extract TogglePaymentsButton; design improvements
