+++ src/pages/popup/components/layout/Header.tsx
@@ -2,11 +2,10 @@ import React from 'react';
 import { Link, useLocation } from 'react-router-dom';
 import { ArrowBack, Settings } from '@/pages/shared/components/Icons';
 import { HeaderEmpty } from './HeaderEmpty';
-import { PowerSwitch } from '@/popup/components/PowerSwitch';
-import { isOkState } from '@/shared/helpers';
+import { TogglePaymentsButton } from '@/popup/components/TogglePaymentsButton';
 import { ROUTES_PATH } from '@/popup/Popup';
-import { useBrowser, useMessage } from '@/popup/lib/context';
-import { usePopupState, dispatch } from '@/popup/lib/store';
+import { useBrowser } from '@/popup/lib/context';
+import { usePopupState } from '@/popup/lib/store';
 
 const NavigationButton = () => {
   const location = useLocation();
@@ -35,28 +34,6 @@ const NavigationButton = () => {
   }, [location, connected]);
 };
 
-export const TogglePaymentsButton = ({
-  large = false,
-  enabled = false,
-}: {
-  large?: boolean;
-  enabled?: boolean;
-}) => {
-  const message = useMessage();
-
-  return (
-    <PowerSwitch
-      enabled={enabled}
-      onChange={() => {
-        message.send('TOGGLE_PAYMENTS');
-        dispatch({ type: 'TOGGLE_PAYMENTS' });
-      }}
-      title="Toggle extension"
-      iconClassName={large ? 'w-32' : 'w-6'}
-    />
-  );
-};
-
 export const Header = () => {
   const browser = useBrowser();
   const { enabled } = usePopupState();
@@ -64,7 +41,7 @@ export const Header = () => {
 
   return (
     <HeaderEmpty logo={Logo}>
-      {enabled && <TogglePaymentsButton enabled={enabled} />}
+      {enabled && <TogglePaymentsButton enabled={true} />}
       <NavigationButton />
     </HeaderEmpty>
   );
