+++ src/pages/popup/pages/Home.tsx
@@ -4,7 +4,7 @@ import { Settings } from '@/pages/shared/components/Icons';
 import { formatNumber, roundWithPrecision } from '@/pages/shared/lib/utils';
 import { PayWebsiteForm } from '@/popup/components/PayWebsiteForm';
 import { NotMonetized } from '@/popup/components/NotMonetized';
-import { TogglePaymentsButton } from '@/popup/components/layout/Header';
+import { TogglePaymentsButton } from '@/popup/components/TogglePaymentsButton';
 import { useTranslation } from '@/popup/lib/context';
 import { usePopupState } from '@/popup/lib/store';
 
@@ -15,8 +15,10 @@ export const Component = () => {
   if (!enabled) {
     return (
       <div className="flex h-full flex-col items-center justify-center gap-4 text-center">
-        <TogglePaymentsButton large={true} />
-        <p>{t('app_text_disabled')}</p>
+        <TogglePaymentsButton enabled={false} large={true} />
+        <p className="text-center text-lg leading-tight text-medium">
+          {t('app_text_disabled')}
+        </p>
       </div>
     );
   }
